requirements.txt
pycairo==1.25.1 